backend/requirements.txt
fastapi
uvicorn
langgraph
google-generativeai
python-dotenv
httpx
pydantic
langchain
langchain-google-genai
langchain-core
